Dateien hochladen nach „http/php“
New CKAN export handler for the rhineland-palatinate transparency portal. This handler is only a workaround. Normally the sync is done by mapbenders ckan classes. If s.o. want to use it, it should be adopted. It uses mapbenders searchInterface to pull all metadata for a special organisation (by using paging).