Add files via upload
Script to qualify layer names which where generated from layer titles when registrating wms. This was done since 2007. We have to distinguish those layers from layers which have real names in the capabilities documents, cause only they may be invoked by getmap and getfeatureinfo operations.  The script may be invoked at a time when not many users are logged in. It compares the layer names from the stored capabilities xml with those from the database.